Move MockHttpResponse to conftest.py
